requirements.txt
streamlit
pandas
scikit-learn
joblib
psycopg2-binary